requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
scrapling>=0.2.0
boto3>=1.34.0
openpyxl>=3.1.2
pandas>=2.2.0
python-dotenv>=1.0.0
Pillow>=10.2.0
tqdm>=4.66.0
